not complet

+++ app/Http/Controllers/Backend/CategoryController.php
@@ -38,6 +38,7 @@ class CategoryController extends Controller
                 }
 
                 $category->save();
+                toastr()->success('Successfully Created!');
                 return redirect()->back();
             }
         }
@@ -57,13 +58,60 @@ class CategoryController extends Controller
 
     public function deleteCategory ($id)
     {
-        $category = Category::find($id);
+        if(Auth::user()){
+            if(Auth::user()->role == 1){
+                $category = Category::find($id);
 
-        if($category->image && file_exists('backend/images/category/'.$category->image)){
-            unlink('backend/images/category/'.$category->image);
+                if($category->image && file_exists('backend/images/category/'.$category->image)){
+                    unlink('backend/images/category/'.$category->image);
+                }
+        
+                $category->delete();
+                toastr()->success('Successfully Deleted!');
+                return redirect()->back();
+            }
         }
+    }
 
-        $category->delete();
-        return redirect()->back();
+    public function editCategory ($id)
+    {
+        if(Auth::user()){
+            if(Auth::user()->role == 1){
+                $category = Category::find($id);
+        
+                return view ('backend.admin.category.edit',compact('category'));
+            }
+        }
+    }
+
+
+    public function updateCategory (Request $request, $id)
+    {
+       if(Auth::user()){
+        if(Auth::user()->role == 1){
+            $category = Category::find($id);
+
+            $category->name = $request->name;
+            $category->slug = Str::slug($request->name);
+    
+            if(isset($request->image)){
+    
+                if($category->image && file_exists('backend/images/category/'.$category->image)){
+                    unlink('backend/images/category/'.$category->image);
+                }
+        
+                $imageName = rand().'-categoryup-'.'.'.$request->image->extension();
+                $request->image->move('backend/images/category/',$imageName);
+    
+                $category->image = $imageName;
+    
+            }
+    
+            $category->save();
+            toastr()->success('Successfully Updated!');
+            return redirect ('/admin/category/list');
+        }
+       }
     }
 }
+
